Add the exec function in templates
This function takes a command as argument, run it inside the sources
tree and return the output. The version option which was handled as a
special case in the source code to execute the command defined in
version_command if it is defined is now a normal option thanks to the
exec function.