Remove the remote/get and remote/put options
The only thing required to define a remote build host now is the
remote/exec option. Instead of using the remote/put and remote/get
options to upload and download files, we are now using tar and the
remote/exec command.

The scripts used to upload and download files are defined in the
remote_put and remote_get options.