Store created input_files to output directory
It used to be created in each project's directory, but it is better to
have them in output directory, to remove them from 'git status'.

For 'project' input_files, the files are stored in the output_dir of
that project.

If you want to store some file in a different directory, it is possible
to override the 'output_dir' option in the input_file definition.