Add the input_files_id option
The value of this option is an identifier of the input_files.
When any of the input files is changed, the identifier changes.
This identifier is something that can be used in a project's
filename to trigger a rebuild when any of its input files is
changed. This identifier is based on: the +input_file_id+ option
of an input file if it is present, the +filename+ for an input
file of type +project+, and the +filename+ and the sha256sum of
the file for any other type of input file.