Avoid modifying $input_files
If an input_file contains a 'target' value we process it as a template
and modify it. However we should not do that as a project (and its
input files) can be used multiple times with different contexts in the
same build. Instead we should make a copy of the $input_file before we
modify it.