Only run 'git checkout --detach' when necessary
Before updating local branches, we need to detach HEAD to be able to
update the current branch. However, we only need to do that if we are
on a branch, and when this branch exists.

This avoids an error when the repository we have cloned does not have a
master branch: in that case we are in a branch that does not exist,
which cause an error when running 'git checkout --detach'.