Bug 23513: run remote_finish when there was an error in remote_start
We were looking at the $remote_tmp_src and $remote_tmp_dst variables to
decide if running remote_finish is needed. However, those variables are
not defined if there was an error in remote_start, and we should still
run remote_finish in that case. Instead we look at the remote_exec
option to decide if we need to run remote_finish.