Bug 23557: download again an input file not matching expected sha256sum
When an input file is already present (downloaded during a previous
build), but is not matching the expected sha256sum, we download it
again.

To do that we rename the input_file_id_need_dl function to
input_file_need_dl, and make it work for all types of actions (instead
of only 'input_files_id' before). In input_file_need_dl we ignore an
the file if it is not matching the expected sha256sum.