Bug 24361: use Path::Tiny instead of File::Slurp
When a build script contains some wide character, rbm fails with an
error:
  Wide character in syswrite at /usr/share/perl5/File/Slurp.pm line 506.

This is because write_file from File::Slurp which we use to write the
build scripts is encoding files to latin-1 by default. While fixing this,
switching to Path::Tiny instead of File::Slurp seems like a good idea:
http://blogs.perl.org/users/leon_timmermans/2015/08/fileslurp-is-broken-and-wrong.html

Switching to Path::Tiny is also improving performance:

  When using File::Slurp:

    $ time ./rbm/rbm showconf --target alpha --target torbrowser-linux-x86_64 tor-browser var/build_id
    82b8cb

    real    0m28.543s
    user    0m24.448s
    sys     0m3.956s

  When using Path::Tiny:

    $ time ~/rbm/rbm showconf --target alpha --target torbrowser-linux-x86_64 tor-browser var/build_id
    82b8cb

    real    0m15.745s
    user    0m13.940s
    sys     0m1.940s

The path function in lib/RBM.pm is conflicting with the path constructor
from Path::Tiny, so we rename it to rbm_path.