Bug 27265: fix output_dir value in sub-projects
In the build_pkg call in input_files, we were setting the output_dir
option to the project's directory, causing this directory to be
incorrectly used as the output directory in some cases. But there is no
need to set output_dir at this point, so we can just remove it.

We also introduce a new feature allowing to set options for the current
project but not for the child projects listed in `input_files`. By
default the options are applied recursively to all projects. If some
options are defined under `norec/` then they are only applied on the
current project.

We use this feature in two places where we were setting the output_dir,
which removes the need to reset it in two places.

At the same time, we fix the `urlget()` call in input_files which was
incorrectly ignoring options.