Bug 28396: Create tmp files in rbm_tmp_dir instead of tmp_dir
rbm_tmp_dir is a directory inside tmp_dir that is automatically removed
when rbm exits.

We also fix an issue in urlget when used within the exec script of an
input file. In this case the input_files_id was depending on the value
of tmp_dir because it is used in urlget. We fix that be setting
getting_input_files_id when computing input_files ids, so that urlget
can avoid using tmp_dir in this case.