Bug 33289: Optimize the get_target function
We also change the behavior of the function: we stop looking after the
first definition of a target. This means for example that if a target is
defined as an array in rbm.conf, and also defined as a hash in the
project's config, then the definition from rbm.conf will be ignored.