Bug 40025: Use hard link to prepare input_files
When using remote_exec, we collect input_files in a temporary directory,
before copying them to the "remote" (for example a container). As we
don't normally modify the files inside this temporary directory, it is
safe to use hard link rather than copies of the files.

When remote_exec is not used, we don't use hard links by default, but
link_input_files can be set to 1 to use hard links.