Bug 40028: Remove p variable from templates
In templates, the `p` variable allowed to access the current project's
rbm config. With the support for projects in modules, it would need to
be adapted. However we never used this variable in tor-browser-build,
and it is not very useful (it's usually better to use c() to access
config), so we remove it.