Bug 40062: Copy input_files directories recursively
When an input_files is a directory, instead of passing each files
included in the directory individually to remote_put, we just pass the
directory and expect remote_put to copy it recursively.

Since in the case of container, each call to remote_put is a significant
overhead, this makes it much faster when including a directory
containing many files.