Bug 40062: Avoid using run_chroot in copy_file_to
When copying files to a container directory, we were running `chown` on
the files using run_chroot, which in addition to using the chroot
command is doing other things like setting up mount points and other
things not needed for running `chown`. Simply using `chroot` should be
enough and faster.