Bug 40072: Move capture_exec to a separate module
Because the RBM and RBM::DefaultConfig modules have circular dependencies,
exporting functions from RBM to RBM::DefaultConfig doesn't work, which
means that part of the changes for rbm#40068 didn't work. We could fix
that by using `RBM::capture_exec` instead of `capture_exec`. An other
solution is to make a separate module for `capture_exec`. Since the RBM
modules is becoming big, I think it's a good idea to move independent
parts to a separate modules.