Bug 40079: make `fetch: if_needed` fetch existing branches
Currently, `fetch: if_needed` will avoid doing a fetch if `git_hash` can
be resolved to a valid git revision. If `git_hash` is a hash or a tag,
there is no need to do a fetch, since those are not supposed to be
changing after a fetch. However, a branch is expected to change, so
doing a fetch in this case is useful.